chore: rename middlewares to middleware
