+++ server/src/routes/municipality.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express'
 import { MunicipalityQuerySchema } from '@/scripts/openapi/docs/api.schema.js'
-import { validate } from '@/middlewares/validate.js'
+import { validate } from '@/middleware/validate.js'
 
 import {
   getMunicipalities,
